15211 - Decoupled the Daily Data user flow smoke tests from the Daily Data page tests.
